ramlat: use the unsigned size_t instead of signed off_t for the random pointer
Gcc 4.9 on ARM over-optimizes the walk operations, it considers that
rnd cannot overflow so it ignores the stop conditions. While this can
be disabled using -fwrapv or -fno-strict-overflow, it's not the
correct method.

The real reason is that rnd is declared as a signed integer, and that
the operation is definitely wrong. It only works because the mask has
the sign bit cleared. Declaring rnd as size_t instead fixes the issue.

+++ ramlat.c
@@ -212,7 +212,7 @@ unsigned int random_read_over_area(void *area, unsigned int usec, size_t size, s
 	size_t mask;
 	unsigned int rounds, loop;
 	uint64_t before, after;
-	off_t rnd = 0;
+	size_t rnd = 0;
 
 	mask = mask_rounded_down(size);
 	mask &= -(size_t)4096;
